fix(category): make in memory service return a pre-configured category (#487)
This sets the category returned from the in-memory driver to a single randomly generated category, regardless of the `id` passed as a rest parameter. Eventually, the driver will support multiple categories and `id`s, but this is a temporary solution that allows for quick UI testing until further support for multiple categories is complete.