build(all): do not publish Ivy libraries to npm (#1250)
Angular recommends not publishing Ivy built libraries yet, so we have to make sure that when we do a build, we don't accidentally publish the Ivy libraries. To do so, we've changed the tsconfigs of our packages to have a "prod" version that disables ivy. This ensures that we don't publish the package to npm as an Ivy package, while still allowing us to work with Ivy internally.