feat(contact): strengthen contact driver types
Previously, the DaffContactServiceInterface was extremely weak in both request and response.

This hardens the type, ironing down the possible send arguments and elimanating the genericity of the response type.

BREAKING CHANGE: The contact driver type changed, if you're providin an implementation, please see the new DaffContactServiceInterface. If you had a use-case for a generic response type, we'd love to hear your feedback.