ci: add Github workflow to lint packages (#2418)
This also adds a few supplemental actions called `clear-old-nx-cache` and
`setup-node` which, respectively, take care of keeping the `nx-cache`
trim and setting up node.

`clear-old-nx-cache` accepts `size-in-mb` param, which allows a consumer
to specify what they want as the maximal storage size of cache.
This is useful because the `lint` cache is
significantly smaller than the `build` cache.

`setup-node` does a little bit more than the "out-of-box" `setup-node`.
It also adds caching and, in particular, a "stamp" cache of "node_modules".
The "stamp" cache is very useful when you don't want skip "npm ci"
or related commands. Stamp-cache is a little dangerous, as the folders
within node_modules can by changed by builds, but it's much faster
(90s vs. 5s).

Co-authored-by: Hanmin Kim <hanminkim123@gmail.com>