ci: add Vercel Deploy from Github actions (#2420)
I've leveraged the pre-existing actions in https://github.com/graycoreio/github-actions
along with a new re-usable workflow (deploy.yml) that can be repeated
for other kinds of environments (say, for example, the demo).

Additionally, I've published the `daffio-16.x` artifact alongside
builds as well. This artifact is zipped ahead-of-time for performance.

Co-authored-by: Hanmin Kim <hanminkim123@gmail.com>