ci: allow fork builds from only trusted contributors
We only allow workflow runs from forks that we know are safe.

We enabled, and dont plan to change from:
    "Require approval for all outside collaborators"
It's a little painful to have to keep clicking the button,
but it's generally safer for us, while still allowing trusted
users to contribute to the project without annoying CI failures.

We have accepted this risk and attempted to handle the concern outlined here:
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/