ci: split apart fork PR previews into workflow_run
We only allow workflow runs from forks that we know are safe via a control
in the Github UI called "Require approval for all outside collaborators".

It's a little painful to have to keep clicking the button in the UI,
but it's generally safer for us (from a security standpoint),
while still allowing trusted users to contribute
to the project without annoying CI failures.

We have understood the risk and attempted to handle the concern outlined here:
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/