fix(product-composite): fix and extract query param (#2437)
The `DaffProductPageLoadSuccess` runs at
different times in the action stream depending on server side render
or not. This meant that, during SSR, the `composite_configuration` was
feature was never used correctly as the Router state is only set AFTER
the `DaffProductPageLoadSuccess`.

Additionally, previously the query param service retrieval was inside
of the effect that handled the query param. This made testing the effect
difficult so I extracted it.