ci: create deployment manually
A few months ago, as a result of security recommendations from Github,
we adjusted PR Previews to be deployed via Workflow run.

This had the unintentional side-effect of making these deployments appear
as if they came from the "develop" branch, as `workflow_run` triggers
always run in the "develop" context.

I've done two things:

1. Adjusted the deployment job to now manually create deployments
   as opposed to using the Github native `environment` because environment
   does not support skipping automatic deployment creation:
   https://github.com/orgs/community/discussions/36919

2. Set the sha for the deployment from the calling workflow_run.

This makes the Github UI work as expected.