feat(design)!: allow modal to be closed with ESC key (#2812)
BREAKING CHANGE: Previously, the `DaffModalService` required a `DaffModal` as an arg to `close` (the return type of `open`). This API was probably larger than it should have been, so I trimmed it down. Most consumers won't notice a return type change, but if you do, we can reconsider this change.