feat(external-router)!: rework to use canMatchFn
BREAKING CHANGE: This is a substantial overhaul to the `@daffodil/external-router`. The most important change is the removal of `DaffExternalRouterExistenceGuard` and supporting services in favor of `daffExternalMatcherTypeGuard`. See the docs for exactly how to change your `Routes`.

Additionally, all tokens have been documented along with new guides for the drivers.

Finally, `daffSeoData` is also available on matched routes. This can be used for adjusting titles and schema data.