feat(daffio): reconfigure how docs are loaded for ssr in various environments
Previously, both @xelaint and I have errors at run time while working locally (and at serve when using a local ssr dev server) as a result of various usages of process.cwd().

Now, we no longer rely on process.cwd() instead, we determine whether or not we're using the Angular dev server and decide how to long assets from there. In addition, when running locally with a ssr server (not vite) we know grab the assets from the built browser folder.